Initial integration of incrementation script with makefile
Assume a pre-populated directory with latest version 0.0.1,
with source beam.ttl (override this with variables).

At the top-level, run `make increment-patch`. This will call ontospy to
generate markdown files, followed by mkdocs to build HTML pages.
RDF syntaxes are then generated.

There is then a call to the ontoprepare script which works out the
incremented version (in this case the patch component), populates the
directories and generates a new `.htaccess` file pointing to the latest
version.

I have tested that this works running on the scratch web server,
and that the different media types are returned as requested, as well as
HTML documentation.