Add sample ShEx schema for `core.ttl`
My sense is that this will be useful for us for cardinality. ShEx seems
to be pretty compact, it feels very natural, and its "Semanctic Actions" (https://www.w3.org/2013/ShEx/Examples/#semanticActions), which are
essentially calls to external functionality during validation, are
likely going to be useful for what we may do with SPARQL in our pipeline.

In composing the ShEx schema, I took the opportunity to clarify certain
elements of staging `core.ttl`. As well as typos and indentation, I
removed orphan `rdfs:subClassOf` elements, and I also made the RDF
properties more succint (e.g. shorten `behaviour_undermines` to
`undermines`).