Second iteration in response to meeting today
Most significantly, this incorporates clarifications to the pattern.
Patterns are still concerned with an asset type, and the edge is attributed
with the role type, but the 'link' property is removed. This works a bit
better than before.

I have added an example instance RDF graph (`./ontology/example-instance.ttl`)
which demonstrates attributing these role types, using reification.
RDF reification is about making statements about statements. This represents
one possible way to model patterns.

I have added a general likelihood class which instances will expand on,
and in the example instance RDF graph, I show the old scheme which used
to be defined in `core.ttl`.

I have made a number of other minor clarifications as well, and I have
expanded the section about SKOS to make it clearer why it is relevant to
this modelling exercise.