Add back the 'node' structure renamed as `ThreatRole`
The key bit of misunderstanding which was clarified today was that 'nodes' are primarily a general role (like attacker subnet), which can be additionally narrowed down to specific classes of asset.

If they were just collections of assets, then there would be an argument to just model an edge between threat and asset, and to attribute that with a possible role. Indeed, modelling it in this way means that roles can be more easily re-used, even if it is relatively uncommon for a role to be meaningful outside of a threat.

On discussion, I have renamed 'node' to "threat role". 'Node' is vague and rather confusing, since we are dealing with network data, whereas this only applies to a certain type of structural entity, and rather captures the role within a threat.

Part of this discussion also involved dispelling with patterns at this stage. We do not anticipate modelling these at all.