Add UML-like diagram to modelling exercise documentation
Note this is much closer to the semi-final RDF vocabulary than what we
started with. There was significant iteration between the diagram and the
RDF vocabulary, and the RDF vocabulary will be visualised similarly.

It is nonetheless still helpful to add it in its paired-down form, and I'm
not sure if we should include the initial diagram yet. Expect the diagrams
included and the documentation (and perhaps the vocabulary) to get modified
further.