ontoconsume: Better account for relations between versioning and history
This makes the following assumpions.

First, in the main RDF graph, there is an association between the
versioned IRI of the RDF graph (e.g.
`http://ontology.spyderisk.org/ns/core#10.0.0`) and a version.

We assume that there is one versioning block of staments in this RDF
graph, and actually remove statements relating to earlier verisons.

Following this, we assume that this version is a blank node, associated
with the RDF graph using the `has_version`. There is only one, so there
isn't much point in naming it, but it is also trivial to change this
behaviour later, so that it can be overriden.

This version is assumed to appear in a log of versions. That blank node
has an assertion made about it, `recorded_as`, which links it to its
corresponding appearance in that log of versions. For example,
`http://ontology.spyderisk.org/v/core#10.0.0`. This relationship makes
most sense in this context. If the script is dealing with a log of
versions, it doesn't add triples with the blank node as subject, but its
identifier in the log of versions.

This required some changes to the script, and modifying the semantics
defined in `endpoint.ttl` accordingly. The script is more robust now and
better tested.

So in the former case (vocabulary), we would end up with something like
the following:

```turtle
_:b1469
    a send:Version ;
    send:major_component 10 ;
    send:minor_component 0 ;
    send:patch_component 0 ;
    send:recorded_as <http://ontology.spyderisk.org/v/core#10.0.0> ;
    send:valid_from "2025-02-21"^^xsd:date ;
    send:see_previous <http://ontology.spyderisk.org/v/core#2.0.0> .

<http://ontology.spyderisk.org/ns/core#10.0.0>
    a rdfg:Graph ;
    send:has_version _:b1469 ;
    send:see_history <http://ontology.spyderisk.org/v/core> .
```

Whereas in the latter case (log of versions), we would end up with
something like the following:

```turtle
<http://ontology.spyderisk.org/v/core#10.0.0>
    a send:Version ;
    send:major_component 10 ;
    send:minor_component 0 ;
    send:patch_component 0 ;
    send:recorded_as <http://ontology.spyderisk.org/v/core#10.0.0> ;
    send:valid_from "2025-02-21"^^xsd:date
    send:see_previous <http://ontology.spyderisk.org/v/core#2.0.0> .

<http://ontology.spyderisk.org/ns/core#10.0.0>
    a rdfg:Graph ;
    send:has_version <http://ontology.spyderisk.org/v/core#10.0.0> ;
    send:see_history <http://ontology.spyderisk.org/v/core> .
```