Incorporate ontoconsume into the Makefile
This assumes that ontoprepare.scm and ontoconsume.scm are installed in
PATH without the extension. This removes relative directories here.

This Makefile is fragile, albeit less so than before. It should really be
a script but I wanted to demonstrate it working end-to-end first. I am
not sure what the best approach with the script would be, but it would
probably also be a Guile scheme script like ontoprepare and ontoconsume.

It incorporates the ontoconsume script for injecting versioning, and as
well as injecting versioning into the input RDF/TTL, it appends to the
log. The nice thing about this is that it also correctly incorporates the
previous version into the injected triples, as long as we call the
Makefile using increment-major, increment-minor or increment-patch.