Update core model to remove edge-level attributes
Edge-level attributes are left commented out in source documents for now.
TargetSystem comprises assets and relations between which, so it can be
used as the focal point (starting shape to test) when validating ShEx schema,
like so:

     shexeval ../examples/example_instance.ttl ./core.shex --start http://ontology.spyderisk.org/sh/core#TargetSystemShape --focus http:/ontology.spyderisk.org/p/ex#Laboratory

This further updates the example instance data and the ShEx schema, primarily
to demonstrate use of RDF:type in order to declare instances of things *at*
specific structural entities. For example, deployment of a buggy script on
a firewall should have a different likelihood depending on *which* firewall
the script is on.

The choice between class inheritance and RDF instantiation is unclear here.
Class inheritance might be desirable if we declare broad classes of threats
or controls which are significantly elaborated on. However, instantiation
with RDF:type feels more natural as in the example above.

Another point of interest is that ShEx doesn't really seem to have a notion
of inheritance, which makes sense as it describes the shape of RDF data,
and the semantics of inheritance differ between the RDF schema and OWL. For now,
the solution is to require that all things which inherit likelihood must
include it in the shape expressions. It's not good enough to, for example,
describe all asset properties with the same shape, because they do differ in
relations like 'undermines'.