Redefine asset_relation as RDF/S class
This defines links_from and links_to as RDF/S properties spanning
asset_relation and Asset.

This is sort of a schema for a graph, the idea being that we potentially
wish to say more about these kind of relations than what we can just by
defining RDF domain and range as before.