Update core.ttl
Fixed description of the Likelihood concept. The idea is that the way you represent likelihood depends on how you choose to simulate a causal model. This will depend on your solver, not on the domain you are modelling (although your choice of solver may depend on the purpose to which you put the model, and that may depend on what you are modelling). For example, you may choose to model likelihood using Bayesian probability values between 0 and 1, or a Likert scale (e.g., 'low', 'medium' and 'high'), or a 'scale of nines' in which a non-negative integer is used (the number of nines needed to distinguish from an impossibility).